// gamemode/src/shared/types/vehicules.d.ts
import { IInventoryItem } from "../player/inventory";

export interface IVehiculesCustom {
	model: string | number;
	plate: string;
	plateIndex: number;
	bodyHealth: number;
	engineHealth: number;
	fuelLevel: number;
	dirtLevel: number;
	color1: number;
	color2: number;
	pearlescentColor: number;
	wheelColor: number;
	wheels: number;
	windowTint: number;
	xenonColor: number;
	neonEnabled: [number, number, number, number];
	neonColor: [number, number, number];
	extras: any[];
	tyreSmokeColor: [number, number, number];
	modSpoilers: number;
	modFrontBumper: number;
	modRearBumper: number;
	modSideSkirt: number;
	modExhaust: number;
	modFrame: number;
	modGrille: number;
	modHood: number;
	modFender: number;
	modRightFender: number;
	modRoof: number;
	modEngine: number;
	modBrakes: number;
	modTransmission: number;
	modHorns: number;
	modSuspension: number;
	modArmor: number;
	modTurbo: number;
	modSmokeEnabled: number;
	modXenon: number;
	modFrontWheels: number;
	modBackWheels: number;
	modPlateHolder: number;
	modVanityPlate: number;
	modTrimA: number;
	modOrnaments: number;
	modDashboard: number;
	modDial: number;
	modDoorSpeaker: number;
	modSeats: number;
	modSteeringWheel: number;
	modShifterLeavers: number;
	modAPlate: number;
	modSpeakers: number;
	modTrunk: number;
	modHydrolic: number;
	modEngineBlock: number;
	modAirFilter: number;
	modStruts: number;
	modArchCover: number;
	modAerials: number;
	modTrimB: number;
	modTank: number;
	modWindows: number;
	modLivery: number;
}

export interface IVehicle {
	name: string;
	plate: string;
	customs: IVehiculesCustom;
	isOut: boolean;
	privateGarageId?: number;
	inventory: IInventoryItem[];
}